Update README.md
Workflow description modified in order to make it more understandable, better harmonised with other helpdesks workflows, and fully integrated with the formal process of action / decision from the INSPIRE Registry Control Body. 
Additional changes to the workflow graphic are needed, at graphical and internal level.
To be commented and discussed during the Task 2 status meeting on 23/05/2022.